feat(tests): fix e2e test harness

+++ standalone/portal-api/test/factories/app.ts
@@ -8,21 +8,34 @@ import Redis from 'ioredis';
 import config from 'config';
 
 import { AppModule } from 'src/app.module';
+import { RedisRateLimiter } from '../../src/modules/auth/limiter/redis.rate.limiter';
+import { NoRateLimiter } from '../../src/modules/auth/limiter/no.rate.limiter';
 
-const dbConfig = config.get('database') as any;
+const dbConfig = config.get('db') as any;
+const throttleConfig = config.get('throttle') as any;
+const throttleEnabled = throttleConfig.enabled as boolean;
 
 export class AppFactory {
   private constructor(
     private readonly appInstance: INestApplication,
-    private readonly redis: Redis.Redis,
-  ) {}
+    private readonly redis?: Redis.Redis,
+  ) { }
 
   get instance() {
     return this.appInstance;
   }
 
   static async new() {
-    const redis = await setupRedis();
+
+    if (throttleEnabled) {
+      const redis = await setupRedis();
+      return await this.setupAppFactory(redis);
+    }
+
+    return await this.setupAppFactory();
+  }
+
+  private static async setupAppFactory(redis?: Redis.Redis) {
     const moduleBuilder = Test.createTestingModule({
       imports: [
         AppModule,
@@ -40,13 +53,17 @@ export class AppFactory {
       .overrideProvider('LOGIN_THROTTLE')
       .useFactory({
         factory: () => {
-          return new RateLimiterRedis({
-            storeClient: redis,
-            keyPrefix: 'login',
-            points: 5,
-            duration: 60 * 60 * 24 * 30, // Store number for 30 days since first fail
-            blockDuration: 3000,
-          });
+          if (redis) {
+            return new RedisRateLimiter(new RateLimiterRedis({
+              storeClient: redis,
+              keyPrefix: 'login',
+              points: 5,
+              duration: 60 * 60 * 24 * 30, // Store number for 30 days since first fail
+              blockDuration: 3000,
+            }));
+          } else {
+            return new NoRateLimiter();
+          }
         },
       });
 
@@ -58,12 +75,32 @@ export class AppFactory {
 
     await app.init();
 
+    const connection = await createConnection({
+      type: dbConfig.type || 'postgres',
+      host: process.env.DB_HOST || dbConfig.host,
+      port: parseInt(process.env.DB_PORT) || dbConfig.port,
+      database: process.env.DB_DATABASE_NAME || dbConfig.database,
+      username: process.env.DB_USERNAME || dbConfig.username,
+      password: process.env.DB_PASSWORD || dbConfig.password,
+      entities: [
+        "src\\**\\*.entity{.ts,.js}"
+      ],
+      migrations: [
+        "src\\database\\migrations\\**\\*{.ts,.js}"
+      ],
+      name: "testing"
+    });
+
+    await connection.synchronize(false); // Drops and creates the schema
+    await connection.close();
     return new AppFactory(app, redis);
   }
 
   async close() {
     await getConnection().dropDatabase();
-    if (this.redis) await this.teardown(this.redis);
+    if (this.redis && throttleEnabled) {
+      await this.teardown(this.redis);
+    }
     if (this.appInstance) await this.appInstance.close();
   }
 
@@ -74,7 +111,28 @@ export class AppFactory {
     );
 
     for (const table of tables) {
-      await connection.query(`DELETE FROM ${table};`);
+      const exists = await connection.query(`SELECT EXISTS (
+        SELECT FROM 
+          pg_catalog.pg_tables 
+        WHERE 
+          schemaname != 'pg_catalog' AND 
+          schemaname != 'information_schema' AND 
+          tablename = '${table}'
+      );`)
+
+      let tableExists = false;
+
+      // result is expected to be an array with one object: [{ exists: boolean }]
+      if (exists && exists.length > 0) {
+        tableExists = exists[0].exists;
+      }
+
+      if (tableExists) {
+        try {
+          await connection.query(`DELETE FROM ${table};`);
+        } catch (e) { }
+      } else {
+      }
     }
   }
 
@@ -89,11 +147,15 @@ export class AppFactory {
     });
 
     await connection.query(`SET session_replication_role = 'replica';`);
+
     const tables = connection.entityMetadatas.map(
       (entity) => `"${entity.tableName}"`,
     );
+
     for (const tableName of tables) {
-      await connection.query(`DROP TABLE IF EXISTS ${tableName};`);
+      try {
+        await connection.query(`DROP TABLE IF EXISTS ${tableName};`);
+      } catch (e) { }
     }
 
     await connection.close();
@@ -101,7 +163,7 @@ export class AppFactory {
 
   async teardown(redis: Redis.Redis) {
     return new Promise<void>((resolve) => {
-      redis.quit();
+      redis.disconnect();
       redis.on('end', () => {
         resolve();
       });
@@ -114,6 +176,7 @@ const setupRedis = async () => {
     host: process.env.REDIS_HOST || 'localhost',
     port: parseInt(process.env.REDIS_PORT) || 6379,
   });
+
   await redis.flushall();
   return redis;
 };
